chore: refactor favorite functions and mock in unit tests
