+++ src/Components/ServiceScene/Breakdowns/AddToFiltersButton.test.tsx
@@ -6,6 +6,13 @@ import userEvent from '@testing-library/user-event';
 import { AdHocFiltersVariable, sceneGraph, SceneObject, SceneQueryRunner } from '@grafana/scenes';
 import { FieldValue, LEVEL_VARIABLE_VALUE, VAR_FIELDS, VAR_LABELS, VAR_LEVELS } from 'services/variables';
 import { ServiceScene, ServiceSceneState } from '../ServiceScene';
+jest.mock('services/favorites', () => {
+  return {
+    rerenderFavorites: () => {},
+    addToFavorites: () => {},
+    removeFromFavorites: () => {},
+  };
+});
 
 const scene = { publishEvent(event: BusEvent, bubble?: boolean) {} } as SceneObject;
 
@@ -185,7 +192,7 @@ describe('addToFilters and addAdHocFilter', () => {
     it('identifies indexed labels and uses the appropriate variable', () => {
       const lookupVariable = jest.spyOn(sceneGraph, 'lookupVariable').mockReturnValue(adHocVariable);
       jest.spyOn(sceneGraph, 'getAncestor').mockReturnValue(serviceScene);
-      addToFilters('indexed', 'value', 'include', scene, VAR_LABELS, true);
+      addToFilters('indexed', 'value', 'include', scene, VAR_LABELS);
 
       expect(lookupVariable).toHaveBeenCalledWith(VAR_LABELS, expect.anything());
       expect(adHocVariable.state.filters).toEqual([
@@ -255,7 +262,7 @@ describe('addToFilters and addAdHocFilter', () => {
 
     it('identifies indexed labels and uses the appropriate variable', () => {
       const lookupVariable = jest.spyOn(sceneGraph, 'lookupVariable').mockReturnValue(adHocVariable);
-      addToFilters('indexed', 'value', 'include', scene, 'filters', true);
+      addToFilters('indexed', 'value', 'include', scene, 'filters');
 
       expect(lookupVariable).toHaveBeenCalledWith(VAR_LABELS, expect.anything());
       expect(adHocVariable.state.filters).toEqual([
